fix param type

+++ src/opts.py
@@ -11,7 +11,7 @@ def get_args():
     parser.add_argument("--debug", action="store_true") # TODo remove in the end
 
     parser.add_argument("--input_size", type=int, default=256)
-    parser.add_argument("--output_size", type=float, default=64)
+    parser.add_argument("--output_size", type=int, default=64)
     parser.add_argument("--num_workers", type=int, default=8)
     
     parser.add_argument("--wh_weight", type=float, default=0.1)
